Update README.md: Added Akade.IndexedSet
Algorithms and Data Structures/[Akade.IndexedSet](https://github.com/akade/Akade.IndexedSet)

Akade.IndexedSet is a library that provides a convenient data structure supporting efficient in-memory indexing and querying, including range queries and fuzzy string matching.

The library is MIT-licensed, is actively developed by me and distributed via NuGet. While it is early in its developement, I think that it is now mature enough to provide actual value to devs. And to the best of my knowledge, I do not know of a similar library in our C#/.Net ecosystem...

Thanks for your consideration. :)